Initial commit for implementing lambdas with the kbv api
