+++ build.gradle
@@ -7,6 +7,7 @@ plugins {
 	id "io.swagger.core.v3.swagger-gradle-plugin" version "2.1.11"
 	id 'com.github.bjornvester.wsdl2java' version '1.2'
 	id "org.sonarqube" version "3.3"
+	id "io.freefair.aspectj.post-compile-weaving" version "6.3.0"
 }
 
 group "uk.gov.di.cri.experian.kbv.api"
@@ -18,13 +19,14 @@ repositories {
 sourceCompatibility = 11
 
 dependencies {
-	implementation 'org.apache.cxf:cxf-rt-frontend-jaxws:3.4.5'
-	implementation 'org.apache.cxf:cxf-rt-transports-http:3.4.5'
-	implementation 'javax.annotation:javax.annotation-api:1.3.2'
-	implementation 'org.slf4j:slf4j-api:1.7.32'
-	implementation 'org.slf4j:slf4j-simple:1.7.32'
-	implementation 'com.sun.xml.messaging.saaj:saaj-impl:1.5.3'
-	implementation "com.sparkjava:spark-core:2.9.3",
+	implementation 'org.apache.cxf:cxf-rt-frontend-jaxws:3.4.5',
+			'org.apache.cxf:cxf-rt-transports-http:3.4.5',
+			'javax.annotation:javax.annotation-api:1.3.2',
+			'org.slf4j:slf4j-api:1.7.32',
+			'org.slf4j:slf4j-simple:1.7.32',
+			'com.sun.xml.messaging.saaj:saaj-impl:1.5.3',
+			"com.amazonaws:aws-lambda-java-core:1.2.1",
+			"com.amazonaws:aws-lambda-java-events:3.11.0",
 			"com.fasterxml.jackson.core:jackson-databind:2.13.0",
 			"com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.13.0",
 			"org.hibernate.validator:hibernate-validator:6.2.1.Final",
@@ -32,6 +34,8 @@ dependencies {
 			"javax.validation:validation-api:2.0.1.Final",
 			"javax.el:javax.el-api:3.0.0",
 			"org.glassfish:javax.el:3.0.0"
+
+	aspect "software.amazon.lambda:powertools-tracing:1.10.2"
 
 	testImplementation "org.junit.jupiter:junit-jupiter:5.8.2",
 			"org.mockito:mockito-junit-jupiter:4.1.0",
