+++ src/main/java/uk/gov/di/cri/experian/kbv/api/App.java
@@ -1,9 +1,9 @@
-package uk.gov.di.cri.experian.kbv.api;
-
-import java.io.IOException;
-
-public class App {
-    public static void main(String[] args) throws IOException {
-        new KBVApi();
-    }
-}
+// package uk.gov.di.cri.experian.kbv.api;
+//
+// import java.io.IOException;
+//
+// public class App {
+//    public static void main(String[] args) throws IOException {
+//        new KBVApi();
+//    }
+// }
