+++ src/main/java/uk/gov/di/cri/experian/kbv/api/domain/Question.java
@@ -1,0 +1,3 @@
+package uk.gov.di.cri.experian.kbv.api.domain;
+
+public class Question {}
